requirements.txt
pyyaml
httpx
tqdm
pillow
pillow-heif
